+++ packages/sketch/src/lib/components/sketch/Sketch.svelte
@@ -45,7 +45,7 @@
 	let id = $derived(meta?.id ? `sketch-${meta.id}` : 'sketch')
 	let debug = dev
 	let filters: Filters = $state(DEFAULT_FILTERS)
-	let canvas: HTMLCanvasElement | null = $state(null)
+	let canvas: HTMLCanvasElement | undefined = $state(undefined)
 	let sceneContext: SceneContext = $state({})
 	let title = meta.title
 	let asset = meta.asset
@@ -54,6 +54,12 @@
 	let frame: number
 	let time: number
 	let resetEvent = $state(0)
+
+	// Snapshot related variables
+	let blobLink: HTMLAnchorElement | undefined = $state(undefined)
+	let blobUrl: string | undefined = $state(undefined)
+	let blobName: string | undefined = $state(undefined)
+	let downloadSnap = $derived(actor.getCurrentEvent() === PlayerEvent.snap)
 
 	let currentAsset = $derived(
 		actor.state.canvas === CanvasState.idle && asset
@@ -194,9 +200,11 @@
 				pause()
 				break
 			case 'clear':
+				clearBlob()
 				clear()
 				break
 			case 'stop':
+				clearBlob()
 				stop()
 				break
 		}
@@ -227,6 +235,74 @@
 		}
 	}
 
+	function takeSnapshot() {
+		clearBlob()
+
+		if (canvas) {
+			canvas.toBlob(
+				(blob: Blob | null) => {
+					if (blob && canvas) {
+						saveBlob()(blob, canvas)
+					}
+				},
+				'image/jpeg',
+				1,
+			)
+		}
+		actor.update(PlayerEvent.snap)
+		downloadSnap = true // TODO: this should be handled by the actor
+	}
+
+	function saveBlob() {
+		// This will timestamp the snapshot with the current date and time (local timezone)
+		// TODO:  make intl helpers for L10n (mote to @fat-fuzzy/intl	...)
+		// https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Intl/DateTimeFormat/formatToParts
+		const date = Date.now()
+		const options = {
+			year: 'numeric',
+			month: 'numeric',
+			day: 'numeric',
+			hour: 'numeric',
+			minute: 'numeric',
+			second: 'numeric',
+			fractionalSecondDigits: 3,
+		}
+		// @ts-expect-error options are OK
+		const dateParts = new Intl.DateTimeFormat([], options)
+			.formatToParts(date)
+			.filter((part) => part.type !== 'literal')
+		const dateString = dateParts
+			.splice(0, 3) // year, month, day
+			.reverse()
+			.map((part) => part.value)
+			.join('-')
+		// const dateTime = dateParts // hour, minute, second, fractionalSecondDigits
+		// 	.map((part, i) => {
+		// 		return i === 0
+		// 			? `h${part.value}`
+		// 			: i === 1
+		// 				? `m${part.value}`
+		// 				: i === 2
+		// 					? `s${part.value}`
+		// 					: `ms${part.value}`
+		// 	})
+		// 	.join('')
+
+		return function saveData(blob: Blob, canvas: HTMLCanvasElement) {
+			blobUrl = URL.createObjectURL(blob)
+			blobName = `snap-${dateString}-${date}-${canvas.width}x${canvas.height}`
+		}
+	}
+
+	function clearBlob() {
+		if (blobUrl) {
+			// If a snapshot was taken, revoke the URL to free memory
+			URL.revokeObjectURL(blobUrl)
+			blobUrl = undefined
+			blobName = undefined
+		}
+	}
+
 	onMount(async () => {
 		try {
 			await init()
@@ -239,6 +315,7 @@
 	onDestroy(() => {
 		try {
 			if (scene) stop()
+			clearBlob()
 		} catch (e: unknown) {
 			actor.feedback.sketch.push({status: 'error', message: e as string})
 			actor.update(SketchEvent.exitNok)
@@ -302,13 +379,15 @@
 	{/snippet}
 
 	{#snippet aside()}
-		<div class="maki:block w:full raviolink">
+		<div class="maki:block l:stack:xs w:full raviolink">
 			{#if canvas}
 				<Player
+					{canvas}
 					play={updateCanvas}
 					pause={updateCanvas}
 					clear={updateCanvas}
 					stop={updateCanvas}
+					snap={takeSnapshot}
 					initial={actor.getPlayButtonState()}
 					{color}
 					size="2xs"
@@ -317,6 +396,18 @@
 					disabled={actor.hasError() ?? undefined}
 					{init}
 				/>
+				<div class="w:full text:center">
+					{#if blobUrl && blobName && downloadSnap}
+						<a
+							bind:this={blobLink}
+							href={blobUrl}
+							download={blobName}
+							class="font:xs raviolink"
+						>
+							Download snapshot
+						</a>
+					{/if}
+				</div>
 				{#if meta && actor.getState('sketch') === 'active' && actor.getIsInteractive()}
 					{#if sceneContext.geometry && meta.controls.includes('matrix-2d')}
 						<Geometry2D