+++ packages/validation/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @fat-fuzzy/validation
 
+## 0.1.8
+
+### Patch Changes
+
+- Enable snapshot, add new Wings, cleanup
+
 ## 0.1.7
 
 ### Patch Changes